--- src/code/compatibility.edge.js
@@ -1,9 +1,0 @@
-window.browser.commands = {
-	onCommand: {
-		addListener: function() {
-			console.log('Landmarks: stub for browser.commands used.')
-		}
-	}
-}
-
-NodeList.prototype.forEach = Array.prototype.forEach
