--- src/code/specialPages.chrome.js
@@ -1,5 +1,0 @@
-const specialPages = Object.freeze([
-	/^https:\/\/chrome.google.com\/webstore/
-])
-
-export default specialPages
