--- src/code/specialPages.edge.js
@@ -1,5 +1,0 @@
-const specialPages = Object.freeze([
-	/^https:\/\/www.microsoft.com\/*?\/store/
-])
-
-export default specialPages
