--- src/code/specialPages.firefox.js
@@ -1,5 +1,0 @@
-const specialPages = Object.freeze([
-	/^https:\/\/addons.mozilla.org/
-])
-
-export default specialPages
