--- src/code/specialPages.opera.js
@@ -1,5 +1,0 @@
-const specialPages = Object.freeze([
-	/^https:\/\/addons.opera.com/
-])
-
-export default specialPages
